pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmap"
version = "0.1.0"
description = "Generate navigable code maps to help LLMs understand project architecture"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Yonder" }
]
keywords = ["llm", "code-analysis", "documentation", "architecture"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "tree-sitter>=0.21",
    "tree-sitter-cpp>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
llmap = "llmap.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["llmap"]
